debug : change int to uint64_t

+++ FINAL_PROJECT/qm_mpi.cpp
@@ -39,8 +39,9 @@ using std::vector;
 namespace mpi = boost::mpi;
 
 // mpic++ qm_mpi.cpp  -std=c++11 -lboost_mpi -lboost_serialization -lmpi  -o
-// qm_mpi  -O3 -g mpic++ qm_mpi.cpp  -std=c++11 -lboost_mpi-mt
-// -lboost_serialization-mt  -o qm_mpi  -O3 -g -L${BOOST_LIB} -I${BOOST_INCLUDE}
+// qm_mpi  -O3 -g
+// mpic++ qm_mpi.cpp  -std=c++11 -lboost_mpi-mt -lboost_serialization-mt  -o
+// qm_mpi  -O3 -g -L${BOOST_LIB} -I${BOOST_INCLUDE}
 
 double time1{0}, time2{0};
 
@@ -103,15 +104,15 @@ struct compareprime {
 };
 
 template <const int Base>
-int convertStrToNum(const string& input) {
-  int res{0};
+uint64_t convertStrToNum(const string& input) {
+  uint64_t res{0};
   for (int i = input.size() - 1, mul = 1; i >= 0; --i, mul *= Base)
     res = res + (input[i] - '0') * mul;
   return res;
 }
 
 template <const int Base>
-string convertNumToStr(int num, int in_bit_num) {
+string convertNumToStr(uint64_t num, int in_bit_num) {
   string res(in_bit_num, '0');
   for (int p = in_bit_num - 1; num; num /= Base) res[p--] = (num % Base) + '0';
   return res;
@@ -275,7 +276,7 @@ void find_primes(mpi::communicator& cmm, vector<string>& v,
     }
   }
 
-  vector<int> vec_primes_local, vec_primes_all;
+  vector<uint64_t> vec_primes_local, vec_primes_all;
   std::transform(prime.begin(), prime.end(),
                  std::back_inserter(vec_primes_local), convertStrToNum<3>);
